Desarrollo diego

+++ Application/Common/Dtos/Coordinates/CoordinatesRequestDto.cs
@@ -9,8 +9,8 @@ namespace Application.Common.Dtos.Coordinates
 {
     public class CoordinatesRequestDto
     {
-        public decimal Longitude { get; set; }
+        public double Longitude { get; set; }
 
-        public decimal Latitude { get; set; }
+        public double Latitude { get; set; }
     }
 }
